requirements.txt
requests
python-dotenv
urllib3<2

# 异步HTTP请求
aiohttp>=3.8.0

# 文档处理
# CentOS 7 gcc 4.8 不支持 C++20，1.23.x 以上需要 C++20，锁定旧版
pymupdf==1.22.5
python-docx>=0.8.11
pdf2image>=1.16.0

# 报表与数据分析
pandas>=1.5.0
openpyxl>=3.0.0

# OCR服务依赖
pytesseract>=0.3.10
Pillow>=9.0.0

# 阿里云OCR (可选)
aliyun-python-sdk-core>=2.13.0
aliyun-python-sdk-ocr>=1.0.0

# 腾讯云OCR (可选)
tencentcloud-sdk-python>=3.0.0

# 百度OCR (可选)
baidu-aip>=4.16.0

# 异步MySQL驱动
aiomysql>=0.1.1
pymysql>=1.0.0

# 定时任务调度
APScheduler>=3.9.0

# 监控告警
sentry-sdk>=1.14.0
slack-sdk>=3.19.0
psutil>=5.9.0
